Separated components into their own file
