+++ src/rosecommon/include/components/graphics.h
@@ -1,0 +1,8 @@
+#pragma once
+
+struct Graphics {
+    Graphics(uint16_t view) : view_(view) {}
+
+    uint16_t view_;
+};
+
