+++ src/rosecommon/include/components/socketconnector.h
@@ -1,0 +1,10 @@
+#pragma once
+
+class CMapClient;
+
+struct SocketConnector {
+    SocketConnector(CMapClient *client) : client_(client) {}
+
+    CMapClient *client_;
+};
+
